creat_pdf.py
from fpdf import FPDF

class PDF(FPDF):
    def header(self):
        self.set_font("Arial", "B", 14)
        self.cell(0, 10, "SHL Assessment Recommender System - Final Report", ln=True, align="C")
        self.ln(5)

    def chapter_title(self, title):
        self.set_font("Arial", "B", 12)
        self.set_text_color(40, 40, 40)
        self.cell(0, 10, title, ln=True, align="L")
        self.ln(2)

    def chapter_body(self, text):
        self.set_font("Arial", "", 11)
        self.set_text_color(50, 50, 50)
        clean_text = text.replace("’", "'").replace("‘", "'").replace("“", '"').replace("”", '"')
        self.multi_cell(0, 8, clean_text)
        self.ln()

# Initialize PDF
pdf = PDF()
pdf.add_page()

# Section 1: Project Overview
pdf.chapter_title("1. Project Overview")
pdf.chapter_body(
    "This project is a Generative AI-powered recommender system built to match job descriptions or job roles with "
    "the most relevant SHL assessments. It uses semantic search based on Sentence Transformers and cosine similarity."
)

# Section 2: Tools & Technologies
pdf.chapter_title("2. Tools & Technologies Used")
pdf.chapter_body(
    "- Python, Streamlit\n"
    "- SentenceTransformers (all-MiniLM-L6-v2)\n"
    "- FAISS (optional for future speed-up)\n"
    "- fpdf2 for PDF reporting\n"
    "- Scikit-learn for evaluation metrics\n"
    "- Web scraping (BeautifulSoup/requests)"
)

# Section 3: Workflow
pdf.chapter_title("3. Workflow & Architecture")
pdf.chapter_body(
    "1. Scrape SHL assessment data and save as JSON.\n"
    "2. Generate embeddings for all assessments.\n"
    "3. Accept user query input (job role or JD).\n"
    "4. Perform cosine similarity matching.\n"
    "5. Return top 5 relevant assessments with details.\n"
    "6. Evaluate using Recall@5 and MAP@5."
)

# Section 4: Evaluation Metrics
pdf.chapter_title("4. Evaluation Insights")
pdf.chapter_body(
    "Queries like 'manage cash register' matched perfectly to 'Cashier Solution'.\n"
    "Mean Recall@5: 1.000\n"
    "Mean Average Precision@5 (MAP@5): 1.000\n"
    "This confirms the model retrieves highly relevant recommendations in real-world HR use-cases."
)

# Section 5: Streamlit App Features
pdf.chapter_title("5. Streamlit Web App Features")
pdf.chapter_body(
    "- Input box for job role or JD snippet.\n"
    "- Displays top 5 SHL assessment matches.\n"
    "- Each result shows title, URL, duration, test types, and adaptive/remote support.\n"
    "- Visual formatting with clean sections for readability."
)

# Section 6: How to Use (End User Guide)
pdf.chapter_title("6. How to Use the App")
pdf.chapter_body(
    "1. Go to the deployed Streamlit app link.\n"
    "2. Enter your job description or job title.\n"
    "3. View the top assessment recommendations.\n"
    "4. Use the links to explore more on SHL's official product pages."
)

# Section 7: Deployment
pdf.chapter_title("7. Deployment on Streamlit Cloud")
pdf.chapter_body(
    "The app is deployed on Streamlit Cloud. Upload your GitHub repo, connect Streamlit, and it goes live in minutes.\n"
    "Example Link: https://shl-assessment-app.streamlit.app"
)

# Section 8: Business Perspective
pdf.chapter_title("8. Business Perspective")
pdf.chapter_body(
    "- Saves time for HR teams by automating test recommendation.\n"
    "- Improves candidate experience with role-relevant assessments.\n"
    "- Scalable for large job datasets or recruitment campaigns."
)

# Section 9: Future Enhancements
pdf.chapter_title("9. Future Enhancements")
pdf.chapter_body(
    "- Add FAISS for faster vector search.\n"
    "- Auto-refresh scraped catalog regularly.\n"
    "- Visual analytics dashboard for recruiters.\n"
    "- Resume parser to extract input automatically."
)

# Save the PDF
pdf.output("final_report.pdf")
print("✅ Final report 'final_report.pdf' created successfully!")
